backend/requirements.txt
Flask==3.0.3
Flask_Cors==5.0.0
pandas==2.2.3
python-dotenv==1.0.1
Requests==2.32.3
